First/requirements.txt
dnspython>=2.4.0

